Crate systemd

Modules

bus
An interface to work with the dbus message bus.

daemon
High-level interface to the systemd daemon module.

id128
API for working with 128-bit ID values, which are a generalizastion of OSF UUIDs (see man 3 sd-id128 for details

journal
High-level interface to the systemd journal.

Macros

char_or_null!
Given an Option<&str>, either returns a pointer to a const char*, or a NULL pointer if None.

log_with!
Similar to log!(), except it accepts a func argument rather than hard coding ::log::log(), and it doesn't filter on log_enabled!().

sd_journal_log!

sd_try!
An analogue of try!() for systemd FFI calls.

Structs

Error
The error type for I/O operations of the Read, Write, Seek, and associated traits.

Functions

ffi_result
Convert a systemd ffi return value into a Result

Type Definitions

Result