Module systemd::bus

Systemd's interface for working with dbus. Allows both dbus servers and clients to be created

Modules

types

Structs

Bus

BusName

BusRef

Error

InterfaceName
A wrapper which promises it always holds a validated dbus interface name

MemberName

Message

MessageRef

ObjectPath
A wrapper which promises it always holds a valid dbus object path

Type Definitions

Result
Result type for dbus calls that contains errors returned by remote services.