Struct systemd::bus::BusName

pub struct BusName {
    // some fields omitted
}

Methods

impl BusName

fn from_bytes(b: &[u8]) -> Result<&Self, &'static str>

Create a bus name reference from a u8 slice.

Users should be careful to ensure all the following requirements are met:

dbus spec 0.26 requires: unique names start with a ':'. well-known names do not. composed of one or more elemenets seperated by a period '.' all elements must be at least 1 character elements can contain only the ASCII characters '[A-Z][a-z][0-9]_-'. elements part of a unique name may begin with a digit. elements in all other bus names must not begin with a digit. must contain at least 1 '.', and thus at least 2 elements must not begin with '.' must be less than the maximum name length (255)

sd-bus additionally requires nul ('\0') termination of the bus name.

unsafe fn from_bytes_unchecked(b: &[u8]) -> &Self

unsafe fn from_ptr_unchecked<'a>(b: *const c_char) -> &'a Self