Struct systemd::bus::BusRef

pub struct BusRef { /* fields omitted */ }

Methods

impl BusRef

pub unsafe fn from_ptr<'a>(r: *const sd_bus) -> &'a BusRef

pub unsafe fn from_mut_ptr<'a>(r: *mut sd_bus) -> &'a mut BusRef

pub fn as_ptr(&self) -> *mut sd_bus

pub fn fd(&self) -> Result<c_int>

Returns the file descriptor used to communicate from a message bus object. This descriptor can be used with poll(3) or a similar function to wait for I/O events on the specified bus connection object.

This corresponds to sd_bus_get_fd

pub fn events(&self) -> Result<c_int>

Returns the I/O events to wait for, suitable for passing to poll or a similar call. Returns a combination of POLLIN, POLLOUT, ... events.

This corresponds to sd_bus_get_events.

pub fn timeout(&self) -> Result<u64>

Returns the time-out in us to pass to poll() or a similar call when waiting for events on the specified bus connection.

This corresponds to sd_bus_get_timeout.

pub fn process(&mut self) -> Result<Option<Option<Message>>>

Drives the connection between the client and the message bus. Each time it is invoked a single operation is executed.

Returns None if no operations were pending (and thus no operations were processed). Returns Some(None) if progress was made but no message was processed. Returns Some(Message) if a message was processed.

This corresponds to sd_bus_process.

pub fn process_priority(
&mut self,
max_priority: i64
) -> Result<Option<Option<Message>>>

This corresponds to sd_bus_process_priority.

pub fn wait(&mut self, timeout: Option<Duration>) -> Result<bool>

Synchronously waits for pending I/O on this Bus object.

After each invocation of wait(), process() should be invoked to process pending I/O work.

Returns true if any I/O was seen.

This corresponds to sd_bus_wait.

pub fn unique_name(&self) -> Result<&BusName>

Get the unique name (address) of this connection to this Bus.

This corresponds to sd_bus_get_unique_name.

pub fn scope(&self) -> Result<&CStr>

pub fn tid(&self) -> Result<pid_t>

pub fn description(&self) -> Result<&CStr>

pub fn address(&self) -> Result<&CStr>

pub fn n_queued_write(&self) -> Result<u64>

pub fn n_queued_read(&self) -> Result<u64>

pub fn method_call_timeout(&self) -> Result<u64>

pub fn bus_id(&self) -> Result<Id128>

pub fn new_signal(
&mut self,
path: &ObjectPath,
interface: &InterfaceName,
member: &MemberName
) -> Result<Message>

This corresponds to sd_bus_message_new_signal.

pub fn new_method_call(
&mut self,
dest: &BusName,
path: &ObjectPath,
interface: &InterfaceName,
member: &MemberName
) -> Result<Message>

This corresponds to sd_bus_message_new_method_call.

pub fn request_name(&mut self, name: &BusName, flags: u64) -> Result<()>

This blocks. To get async behavior, use request_name_async()

This corresponds to sd_bus_request_name

pub fn request_name_async<F>(
&mut self,
name: &BusName,
flags: u64,
callback: F
) -> Result<()> where
F: Fn(&mut MessageRef) -> Result<()> + Send + Sync + 'static,

pub fn release_name(&self, name: &BusName) -> Result<()>

This blocks. To get async behavior, use request_name directly.

pub fn add_object<F>(&self, path: &ObjectPath, callback: F) -> Result<()> where
F: Fn(&mut MessageRef) -> Result<()> + Send + Sync + 'static,

This corresponds to sd_bus_add_object

pub fn add_object_manager(&self, path: &ObjectPath) -> Result<()>