Struct systemd::bus::InterfaceName

pub struct InterfaceName { /* fields omitted */ }

A wrapper which promises it always holds a validated dbus interface name

Methods

impl InterfaceName

pub fn from_bytes(b: &[u8]) -> Result<&InterfaceName, &'static str>

Create a interface name reference from a u8 slice.

Users should be careful to ensure all the following requirements are met:

dbus spec 0.26 requires: composed of 1 or more elements seperated by a period ('.') character. Elements contain at least 1 character Elements must contain only the ASCII characters '[A-Z][a-z][0-9]_' and must not begin with a digit Interface names must contain at least one '.' character (and thus at least 2 elements) Interface names must not being with a '.' character sd-bus additionally requires nul ('\0') termination of the interface name.

pub unsafe fn from_bytes_unchecked(b: &[u8]) -> &InterfaceName

Unsafety:

b must be a nul terminated string

b must contain a valid interface

pub unsafe fn from_ptr_unchecked<'a>(b: *const c_char) -> &'a Self

Unsafety:

lifetime 'a must be valid

b must be a nul terminated string

b must contain a valid interface