Struct systemd::bus::MemberName

pub struct MemberName { /* fields omitted */ }

Methods

impl MemberName

pub fn from_bytes(b: &[u8]) -> Result<&Self, &'static str>

Create a member name reference from a u8 slice.

Users should be careful to ensure all the following requirements are met:

dbus spec 0.26 requires: must only contain the ASCII characters '[A-Z][a-z][0-9]_' and may not begin with a digit must not contain the '.' character must not exceed the maximum name length (255) must be at least 1 byte in length

sd-bus additionally requires nul ('\0') termination of the bus name.

Important traits for &'a mut W
impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn from_bytes_unchecked(b: &[u8]) -> &Self

Important traits for &'a mut W
impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn from_ptr_unchecked<'a>(b: *const c_char) -> &'a Self