Struct systemd::bus::Message

pub struct Message { /* fields omitted */ }

A message to be sent or that was recieved over dbus

This is reference counted, cloned objects refer to the same root object.

Methods

impl Message

pub unsafe fn take_ptr(p: *mut sd_bus_message) -> Message

Construct a Message, taking over an already existing reference count on the provided pointer

To construct a Message from an un-owned pointer, use MessageRef::from_ptr(p).to_owned()

pub fn as_ptr(&self) -> *const sd_bus_message

pub fn as_mut_ptr(&mut self) -> *mut sd_bus_message