Struct systemd::bus::MessageRef

pub struct MessageRef { /* fields omitted */ }

A reference to a Message

Methods

impl MessageRef

pub fn bus(&self) -> &BusRef

pub fn set_destination(&mut self, dest: &BusName) -> Result<()>

Set the message destination, the name of the bus client we want to send this message to.

XXX: describe broadcast

Fails if the message is sealed

pub fn set_auto_start(&mut self, yes: bool) -> Result<()>

Set to true to allow the bus to launch an owner for the destination name.

Set to false to prevent the bus from launching an owner for the destination name.

Fails if the message is sealed

This controls the NO_AUTO_START dbus header flag.

The specification covers some details about the auto start mechanism, but not all of it is specified.

pub fn send(&mut self) -> Result<u64>

Send expecting a reply. Returns the reply cookie.

Seals self.

pub fn send_no_reply(&mut self) -> Result<()>

Send without expecting any reply Seals self.

pub fn send_to(&mut self, dest: &BusName) -> Result<u64>

Send this message to a destination.

Internally, this is the same as .set_destination() + .send() Seals self.

pub fn send_to_no_reply(&mut self, dest: &BusName) -> Result<()>

Same as self.send_to(), but don't expect a reply. Seals self.

pub fn call(&mut self, usec: u64) -> Result<Message>

Use this message to call a dbus method. Blocks until a reply is recieved or usec microseconds elapse (ie: this times out)

XXX: document blocking forever Seals self.

pub fn call_async<F: FnMut(&mut MessageRef) -> Result<()>>(
&mut self,
callback: &mut F,
usec: u64
) -> Result<()>

Use this message to call a dbus method. Returns immediately and will call the callback when a reply is recieved.

XXX: document how timeout affects this Seals self.

pub fn new_method_error(&mut self, error: &Error) -> Result<Message>

pub fn new_method_return(&mut self) -> Result<Message>

pub unsafe fn append_basic_raw(
&mut self,
dbus_type: u8,
v: *const c_void
) -> Result<()>

Raw access to append data to this message Will fail if the message is sealed

pub fn append<V: ToSdBusMessage>(&mut self, v: V) -> Result<()>

Append a value to the message

pub fn iter<'a>(&'a mut self) -> Result<MessageIter<'a>>

Get an iterator over the message. This iterator really exists with in the Message itself, so we can only hand out one at a time.

Ideally, handing this iterator out wouldn't prevent the use of other non-iterator accessors, but right now it does (unless you bypass borrowck using unsafe{})

Requires that message is sealed.