Trait systemd::bus::types::SdBusMessageDirect

pub unsafe trait SdBusMessageDirect {
    fn dbus_type() -> u8;
}

When impled for a Type, promises that a reference to the type cast to a pointer can be encoded using the given dbus_type() using sd_bus_message_append_basic and decoded similarly using sd_bus_message_read_basic

It is unlikely that this is something you want to impliment, all the basic types are already implimented internally.

NOTE: Ideally, we'd use an associated const here. When those land on stable this may change to that.

Required methods

fn dbus_type() -> u8

Loading content...