Function systemd::daemon::is_socket_inet

pub fn is_socket_inet(
    fd: Fd, 
    family: Option<c_uint>, 
    socktype: Option<SocketType>, 
    listening: Listening, 
    port: Option<u16>
) -> Result<bool>

Identifies whether the passed file descriptor is an Internet socket. If family, type, and/or port are supplied, they must match as well. See Listening for listening check parameters.