Function systemd::daemon::is_socket_unix

pub fn is_socket_unix<S: CStrArgument>(
    fd: Fd, 
    socktype: Option<SocketType>, 
    listening: Listening, 
    path: Option<S>
) -> Result<bool>

Identifies whether the passed file descriptor is an AF_UNIX socket. If type are supplied, it must match as well. For normal sockets, leave the path set to None; otherwise, pass in the full socket path. See Listening for listening check parameters.