Module systemd::journal

High-level interface to the systemd journal.

The main interface for writing to the journal is fn log(), and the main interface for reading the journal is struct Journal.

Structs

Journal
A reader for systemd journal.

JournalLog
Logger implementation over systemd-journald.

Enums

JournalFiles
Represents the set of journal files to read.

JournalSeek
Seeking position in journal.

Functions

log
Record a log entry, with custom priority and location.

log_record
Send a log::LogRecord to systemd-journald.

log_target
Record a log entry, with custom priority, location and target.

print
Send a simple message to systemd-journald.

send
Send preformatted fields to systemd.

Type Definitions

JournalRecord