Module systemd::login

Interface to introspect on seats, sessions and users.

Enums

UnitType
Systemd slice and unit types

Functions

get_cgroup
Determines the control group path of a process.

get_machine_name
Determines the machine name of a process.

get_owner_uid
Determines the owner uid of a process.

get_seat
Determines the seat identifier of the seat the session identified by the specified session identifier belongs to.

get_session
Determines the session identifier of a process.

get_slice
Determines the slice (either in system or user session) of a process.

get_unit
Determines the systemd unit (i.e. service or scope unit) identifier of a process.

get_vt
Determines the VT number of the session identified by the specified session identifier.