Crate systemd

Modules

bus
An interface to work with the dbus message bus.

daemon
High-level interface to the systemd daemon module.

id128
APIs to process and generate 128-bit ID values for systemd.

journal
High-level interface to the systemd journal.

login
Interface to introspect on seats, sessions and users.

Macros

log_with
Similar to log!(), except it accepts a func argument rather than hard coding ::log::log(), and it doesn't filter on log_enabled!().

sd_journal_log

sd_try
An analogue of try!() for systemd FFI calls.

Structs

Error
The error type for I/O operations of the Read, Write, Seek, and associated traits.

Functions

ffi_result
Convert a systemd ffi return value into a Result

Type Definitions

Result
A specialized Result type for I/O operations.