Module systemd::bus

An interface to work with the dbus message bus.

Modules

types
Define a mechanism for converting types to messages and messages to types

Structs

Bus

BusName

BusRef

Error

InterfaceName
A wrapper which promises it always holds a validated dbus interface name

MemberName

Message
A message to be sent or that was recieved over dbus

MessageIter
An iterator over the elements of a Message, use this to read data out of a message.

MessageRef
A reference to a Message

ObjectPath
A wrapper which promises it always holds a valid dbus object path

RawError

Enums

MessageType

Type Definitions

Result
Result type for dbus calls that contains errors returned by remote services (and local errors as well).