Struct systemd::bus::MessageIter

pub struct MessageIter<'a> { /* fields omitted */ }

An iterator over the elements of a Message, use this to read data out of a message.

Note: we're using a concrete type here instead of a reference to allow us to handle lifetimes properly.

Methods

impl<'a> MessageIter<'a>

pub unsafe fn read_basic_raw<R, T, F: FnOnce(R) -> T>(
&mut self,
dbus_type: u8,
cons: F
) -> Result<Option<T>> where
T: 'a,

Read an element from the message and advance the internal cursor References returned by this function are valid until the iterator itself is dropped (just to garuntee they don't change).

XXX: really, they are valid until the message is un-sealed: reading from the message can only occur while the message is sealed. Unclear if we can track lifetimes against message sealing.

Unsafety:

dbus_type when given to sd_bus_message_read_basic() must result in something that can be reinterpreted as R.

pub fn peek_type(&mut self) -> Result<(c_char, &str)>

This needs to be &mut as the &str will be invalid after either of:

self is dropped

sd_bus_message_peek_type is called a second time

Using &mut allows us to prevent #2.

FIXME/WARNING: Message might have been cloned, in which case we can't rely on the lifetime of &str! As Message isn't Send or Sync, we can garuntee we're not racing with someone else to free it though. Probably need to allocate space for it here rather than return a ref.

pub fn next<V: FromSdBusMessage<'a>>(&'a mut self) -> Result<Option<V>>