Struct systemd::bus::ObjectPath

pub struct ObjectPath {
    // some fields omitted
}

A wrapper which promises it always holds a valid dbus object path

Requirements (from dbus spec 0.26):

path must begin with ASCII '/' and consist of elements separated by slash characters

each element must only contain the ASCII characters '[A-Z][a-z][0-9]_'

No element may be the empty string

Multiple '/' characters may not occur in sequence

A trailing '/' character is not allowed unless the path is the root path

Further, sd-bus additionally requires nul ('\0') termination of paths.

Methods

impl ObjectPath

fn from_bytes(b: &[u8]) -> Result<&ObjectPath, &'static str>

Create a path reference from a u8 slice. Performs all checking needed to ensure requirements are met.

unsafe fn from_bytes_unchecked(b: &[u8]) -> &ObjectPath

unsafe fn from_ptr_unchecked<'b>(b: *const c_char) -> &'b ObjectPath