Module systemd::bus::types

Define a mechanism for converting types to messages and messages to types

A few existing models:

rust-postgres: defines ToSql and FromSql traits, and has their accessor and creator functions take slices of trait objects. rust-postgress is not an ffi lib, so they read directly from a Read and write to a Write. For them, each item is associated with a postgres type, so they provide a mechanism in ToSql/FromSql to determine if a particular pg-type can decode to a given rs-type. Essentially, they allow picking a decoding mechanism based on the pg-type.

dbus-rs: defines a MessageItem enum which maps dbus types to rust types. All types are owned (so that they can be decoded into). This will result in higher overhead than allowing users to perform more direct convertions

serde: we could think about our encoding like plain-old serialization/deserialization & provide the serde/rustc-serialize interfaces into it. This is probably more work but is potentially the most convenient.

Structs

UnixFd
A basic wrapper that simply ensures we send a Fd via the dbus file descriptor mechanisms rather than as a integer

Traits

FromSdBusMessage
Allows types to provide a convertion from a dbus message

SdBusMessageDirect
When impled for a Type, promises that a reference to the type cast to a pointer can be encoded using the given dbus_type() using sd_bus_message_append_basic and decoded similarly using sd_bus_message_read_basic

ToSdBusMessage
Allows types to provide a convertion to a dbus message